WIP: added ability for "importing" simulations to delete fields, macros, globals and functions defined in an imported simulation
the trick is to set them to null or ~ (yaml syntaxes for None)

WIP because this still needs a changelog entry and probably some cleanup. I would rather ignore all of them in the same place, not all over the place. So I guess, I will have to do it just after handling imports, like globals.